output: split output module
This splits the output module to keep the texture handling separate.
