+++ src/libtrx/meson.build
@@ -133,7 +133,8 @@ sources = [
   'game/objects/common.c',
   'game/objects/names.c',
   'game/objects/vars.c',
-  'game/output.c',
+  'game/output/common.c',
+  'game/output/textures.c',
   'game/packer.c',
   'game/phase/executor.c',
   'game/phase/phase_cutscene.c',
